improve publish process

+++ Pub/PublishApps.ps1
@@ -1,6 +1,7 @@
 param( 
 	[Parameter(Mandatory=$true)][object]$bump,
 	[Parameter(Mandatory=$true)][object]$nugets,
+	[Parameter(Mandatory=$true)][object]$client,
 	[Parameter(Mandatory=$true)][object]$android, 
 	[Parameter(Mandatory=$true)][object]$server,
 	[Parameter(Mandatory=$true)][object]$distribute
@@ -20,22 +21,28 @@ $noclean = $true;
 $noPushNuget = !$nugets
 
 Remove-Item "$packagesRootDir/ReleaseNote.txt" -ErrorAction Ignore;
-Remove-Item $packagesClientDir -ErrorAction Ignore -Recurse;
 
 # rebuild libraries
 & "$solutionDir\VpnHood.Common\_publish.ps1"
 & "$solutionDir\VpnHood.Tunneling\_publish.ps1"
 
 & "$solutionDir\VpnHood.Client\_publish.ps1"
+& "$solutionDir\VpnHood.Client.Device.Android\_publish.ps1"
 & "$solutionDir\VpnHood.Client.Device\_publish.ps1"
 & "$solutionDir\VpnHood.Client.Device.WinDivert\_publish.ps1"
-& "$solutionDir\VpnHood.Client.Device.Android\_publish.ps1"
-
 & "$solutionDir\VpnHood.Client.App\_publish.ps1"
 & "$solutionDir\VpnHood.Client.App.UI\_publish.ps1"
 
 & "$solutionDir\VpnHood.Server\_publish.ps1"
 & "$solutionDir\VpnHood.Server.Access\_publish.ps1"
+
+# publish client
+if ($client)
+{
+	Remove-Item $packagesClientDir -ErrorAction Ignore -Recurse;
+	& "$solutionDir/VpnHood.Client.App.Win/_publish.ps1";
+	& "$solutionDir/VpnHood.Client.App.Win.Setup/_publish.ps1";
+}
 
 # publish server
 if ($server)
@@ -43,10 +50,6 @@ if ($server)
 	Remove-Item $packagesServerDir -ErrorAction Ignore -Recurse;
 	& "$solutionDir/VpnHood.Server.App.Net/_publish.ps1" -ftp:$false;
 }
-
-# publish client
-& "$solutionDir/VpnHood.Client.App.Win/_publish.ps1";
-& "$solutionDir/VpnHood.Client.App.Win.Setup/_publish.ps1";
 
 if ($android)
 {
